+++ src/yaplox/class_type.py
@@ -1,0 +1,6 @@
+import enum
+
+
+class ClassType(enum.Enum):
+    NONE = enum.auto()
+    CLASS = enum.auto()