+++ src/yaplox/function_type.py
@@ -4,3 +4,5 @@ import enum
 class FunctionType(enum.Enum):
     NONE = enum.auto()
     FUNCTION = enum.auto()
+    INITIALIZER = enum.auto()
+    METHOD = enum.auto()