+++ src/yaplox/parser.py
@@ -5,13 +5,27 @@ from yaplox.expr import (
     Binary,
     Call,
     Expr,
+    Get,
     Grouping,
     Literal,
     Logical,
+    Set,
+    This,
     Unary,
     Variable,
 )
-from yaplox.stmt import Block, Expression, Function, If, Print, Return, Stmt, Var, While
+from yaplox.stmt import (
+    Block,
+    Class,
+    Expression,
+    Function,
+    If,
+    Print,
+    Return,
+    Stmt,
+    Var,
+    While,
+)
 from yaplox.token import Token
 from yaplox.token_type import TokenType
 
@@ -44,6 +58,8 @@ class Parser:
 
     def _declaration(self) -> Optional[Stmt]:
         try:
+            if self._match(TokenType.CLASS):
+                return self._class_declaration()
             if self._match(TokenType.FUN):
                 return self._function("function")
             if self._match(TokenType.VAR):
@@ -53,7 +69,18 @@ class Parser:
             self._synchronize()
             return None
 
-    def _function(self, kind: str) -> Stmt:
+    def _class_declaration(self) -> Stmt:
+        name = self._consume(TokenType.IDENTIFIER, "Expect class name.")
+        self._consume(TokenType.LEFT_BRACE, "Expect '{' before class body.")
+
+        methods = []
+        while not self._check(TokenType.RIGHT_BRACE) and not self._is_at_end():
+            methods.append(self._function("method"))
+
+        self._consume(TokenType.RIGHT_BRACE, "Expect '}' after class body.")
+        return Class(name=name, methods=methods)
+
+    def _function(self, kind: str) -> Function:
         name = self._consume(TokenType.IDENTIFIER, f"Expect {kind} name.")
         self._consume(TokenType.LEFT_PAREN, f"Expect '(' after {kind} name.")
 
@@ -216,6 +243,8 @@ class Parser:
             if isinstance(expr, Variable):
                 name = expr.name
                 return Assign(name=name, value=value)
+            elif isinstance(expr, Get):
+                return Set(expr.obj, expr.name, value)
             self._error(equals, "Invalid assignment target.")
         return expr
 
@@ -314,6 +343,11 @@ class Parser:
         while True:
             if self._match(TokenType.LEFT_PAREN):
                 expr = self._finish_call(expr)
+            elif self._match(TokenType.DOT):
+                name = self._consume(
+                    TokenType.IDENTIFIER, "Expect property name after '.'."
+                )
+                expr = Get(expr, name)
             else:
                 break
         return expr
@@ -330,6 +364,9 @@ class Parser:
 
         if self._match(TokenType.NUMBER, TokenType.STRING):
             return Literal(self._previous().literal)
+
+        if self._match(TokenType.THIS):
+            return This(self._previous())
 
         if self._match(TokenType.IDENTIFIER):
             return Variable(self._previous())