+++ CHANGELOG.md
@@ -9,7 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Completing chapter 12
+- Class can be printed, instantiated, get, set expression, methods on classes, implement `this`, init()
+- YaploxClass created, and YaploxCallable
 - Use [poetry-dynamic-versioning](https://github.com/mtkennerly/poetry-dynamic-versioning) to generate version numbers
+
+### Changed
+
+- In generate_ast.py use `from __future__ import annotations` in generated files
 
 ## [0.0.8] - 2020-10-14